revamp: productOption's bulkaction in BulkActions component
