+++ admin/src/apps/products/views/Listings/ProductsListing/index.jsx
@@ -139,6 +139,7 @@ const ProductsListing = () => {
                         close={closeTunnel}
                         selectedRows={selectedRows}
                         removeSelectedRow={removeSelectedRow}
+                        setSelectedRows={setSelectedRows}
                      />
                   </ModifiersProvider>
                ) : (
@@ -166,7 +167,6 @@ const ProductsListing = () => {
             >
                <Banner id="products-app-products-listing-top" />
                <Flex container alignItems="center">
-
                   <Text as="h2" style={{ marginBottom: 0 }}>
                      {t(address.concat('products'))}
                   </Text>
@@ -249,6 +249,7 @@ class DataTable extends React.Component {
       this.handleMultipleRowSelection =
          this.handleMultipleRowSelection.bind(this)
       this.tableRef = React.createRef()
+      this.optionTableRef = React.createRef()
       this.handleRowSelection = this.handleRowSelection.bind(this)
    }
 
@@ -375,8 +376,8 @@ class DataTable extends React.Component {
       )
       const productGroupParse =
          productGroup !== undefined &&
-            productGroup !== null &&
-            productGroup.length !== 0
+         productGroup !== null &&
+         productGroup.length !== 0
             ? JSON.parse(productGroup)
             : null
       this.tableRef.current.table.setGroupBy(
@@ -465,36 +466,36 @@ class DataTable extends React.Component {
    render() {
       const selectionColumn =
          this.props.selectedRows.length > 0 &&
-            this.props.selectedRows.length < this.props.data.length
+         this.props.selectedRows.length < this.props.data.length
             ? {
-               formatter: 'rowSelection',
-               titleFormatter: reactFormatter(
-                  <CrossBox
-                     removeSelectedProducts={this.removeSelectedProducts}
-                  />
-               ),
-               align: 'center',
-               hozAlign: 'center',
-               width: 10,
-               headerSort: false,
-               frozen: true,
-            }
+                 formatter: 'rowSelection',
+                 titleFormatter: reactFormatter(
+                    <CrossBox
+                       removeSelectedProducts={this.removeSelectedProducts}
+                    />
+                 ),
+                 align: 'center',
+                 hozAlign: 'center',
+                 width: 10,
+                 headerSort: false,
+                 frozen: true,
+              }
             : {
-               formatter: 'rowSelection',
-               titleFormatter: reactFormatter(
-                  <CheckBox
-                     checked={this.state.checked}
-                     handleMultipleRowSelection={
-                        this.handleMultipleRowSelection
-                     }
-                  />
-               ),
-               align: 'center',
-               hozAlign: 'center',
-               width: 20,
-               headerSort: false,
-               frozen: true,
-            }
+                 formatter: 'rowSelection',
+                 titleFormatter: reactFormatter(
+                    <CheckBox
+                       checked={this.state.checked}
+                       handleMultipleRowSelection={
+                          this.handleMultipleRowSelection
+                       }
+                    />
+                 ),
+                 align: 'center',
+                 hozAlign: 'center',
+                 width: 20,
+                 headerSort: false,
+                 frozen: true,
+              }
       return (
          <>
             {this.props.view == 'simple' && (
@@ -536,8 +537,9 @@ class DataTable extends React.Component {
                                  rowSelected={this.handleRowSelection}
                                  rowDeselected={this.handleDeSelection}
                                  options={{
-                                    ...tableOptions, persistenceID: `${this.props.view}_product_table`,
-                                    reactiveData: true
+                                    ...tableOptions,
+                                    persistenceID: `${this.props.view}_product_table`,
+                                    reactiveData: true,
                                  }}
                                  data-custom-attr="test-custom-attribute"
                                  className="custom-css-class"
@@ -587,8 +589,9 @@ class DataTable extends React.Component {
                         rowSelected={this.handleRowSelection}
                         rowDeselected={this.handleDeSelection}
                         options={{
-                           ...tableOptions, persistenceID: `${this.props.view}_product_table`,
-                           reactiveData: true
+                           ...tableOptions,
+                           persistenceID: `${this.props.view}_product_table`,
+                           reactiveData: true,
                         }}
                         data-custom-attr="test-custom-attribute"
                         className="custom-css-class"
@@ -675,8 +678,8 @@ const ActionBar = ({
       )
       const productGroupParse =
          productGroup !== undefined &&
-            productGroup !== null &&
-            productGroup.length !== 0
+         productGroup !== null &&
+         productGroup.length !== 0
             ? JSON.parse(productGroup)
             : null
       if (productGroupParse !== null) {
@@ -718,8 +721,8 @@ const ActionBar = ({
                   {selectedRows.length == 0
                      ? `No ${title}`
                      : selectedRows.length == 1
-                        ? `${selectedRows.length} ${title}`
-                        : `${selectedRows.length} ${title}s`}{' '}
+                     ? `${selectedRows.length} ${title}`
+                     : `${selectedRows.length} ${title}s`}{' '}
                   selected
                </Text>
                <ButtonGroup align="left">
@@ -881,7 +884,7 @@ const ProductOptions = forwardRef(
             setProductOptionsList(newOptions)
          },
       })
-      useEffect(() => { }, [])
+      useEffect(() => {}, [])
       const columns = [
          {
             title: 'Label',
@@ -961,7 +964,6 @@ const ProductOptions = forwardRef(
       }
 
       const handleGroupBy = option => {
-
          tableRef.current.table.setGroupBy(['label', ...option])
       }
       const handleRowSelection = ({ _row }) => {
@@ -971,8 +973,8 @@ const ProductOptions = forwardRef(
          )
          const lastPersistenceParse =
             lastPersistence !== undefined &&
-               lastPersistence !== null &&
-               lastPersistence.length !== 0
+            lastPersistence !== null &&
+            lastPersistence.length !== 0
                ? JSON.parse(lastPersistence)
                : []
          setSelectedRows(prevState => [...prevState, _row.getData()])
@@ -989,8 +991,8 @@ const ProductOptions = forwardRef(
          )
          const lastPersistenceParse =
             lastPersistence !== undefined &&
-               lastPersistence !== null &&
-               lastPersistence.length !== 0
+            lastPersistence !== null &&
+            lastPersistence.length !== 0
                ? JSON.parse(lastPersistence)
                : []
          setSelectedRows(prevState =>
@@ -1016,8 +1018,8 @@ const ProductOptions = forwardRef(
          )
          const productOptionsGroupParse =
             productOptionsGroup !== undefined &&
-               productOptionsGroup !== null &&
-               productOptionsGroup.length !== 0
+            productOptionsGroup !== null &&
+            productOptionsGroup.length !== 0
                ? JSON.parse(productOptionsGroup)
                : null
          tableRef.current.table.setGroupBy(
